Refactor Location implementation.
